Add files via upload
I used AI to help create a Ubuntu pentesting lab with forensics, private ips, 3 users with sudo and regular user account and it should be scalable. The general user should be able to add an extra user and the script should add it to the private network. 

This can be useful for people who want to create pentesting labs or soc labs. It would give you a fresh ubuntu machine with lots of tools and networks automatically setup. 

If anyone has any ideas to add please go head and suggest them.